Put in a few unsigned/signed fixes, etc.
CVS patchset: 2258
CVS date: 2003/12/17 17:33:44


Mercurial revision: 6586394261b7507757fb6e6255265d765b61ff89

+++ telecom/checkpointd/message.h
@@ -122,7 +122,7 @@ typedef struct _SaCkptMessageT {
 	
 	int		paramLength;
 	void*		param;
-	int		dataLength;
+	size_t		dataLength;
 	void*		data;
 
 	SaErrorT	retVal;
